Add couch_access_log module, a common log format logger.
This module introduces a new log file couch_access.log in the
spirit of Apache httpd's access.log.

The log format is specified here:

  http://en.wikipedia.org/wiki/Common_Log_Format

There is currently no way to configure a different log format, but
that can be added easily later.

This is a new module rather than an extension to the couch_log
module to allow easy addition to Apache CouchDB. Ideally, this
will be a plugin in the future.

A note for chunked responses, the response body length reported is
"chunked" rather than the number of bytes. The case is unspecified
in the format description. Maybe the fix is keeping track of bytes
sent in a chunked response, but I didn't want to make the first stab
at this more complicated that it has to be. I'm happy to adjust.

Change-Id: Ie314a861557dfa3eceedd3eb43d8753c6710e5db
Reviewed-on: http://review.couchbase.org/9063
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Tested-by: Filipe David Borba Manana <fdmanana@gmail.com>